improved and extended error handling (discrete ruby errors) + testing
Most of the commands are now run through a private run_command method which checks the exit status and augeas_error. If any error is found, then it is raised as a ruby error. A lot more errors (hopefully all?) are now raised instead of passing silently and being stored in the augeas tree.